requirements.txt
aiohttp==1.0.5
async-timeout==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
discord.py==0.16.12
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
multidict==4.4.2
numpy==1.15.2
opencv-python==3.4.3.18
pyzmq==17.1.2
requests==2.20.0
urllib3==1.23
websockets==3.4
Werkzeug==0.14.1
